refactor(tui): extract chat viewport controller
